trying long long

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,8 +9,8 @@
  */
 int _atoi(char *s)
 {
-	int num;
+	long long num;
 
-	num = atoi(s);
+	num = atoll(s);
 	return (num);
 }
